Preliminary Network code (see description)
With this commit it's possible to access all days of the calendar, but
it's not reliable.

PostRequest and GetRequest are classes that let you POST or GET using
HTTP.

Unfortunately, the CSUSB web server is unreliable and it has a chance of
failing and returning a redirect instead of a real response from a POST.

All that's left is to parse the data and do a final POST on Booking
Details.